Lift: move TH to its own file
This matches the other classes, so the `Class` file just has the classes
and the straightforward instances.
