+++ plutus-tx/plutus-tx.cabal
@@ -89,6 +89,7 @@ library
     PlutusTx.IsData.Instances
     PlutusTx.IsData.TH
     PlutusTx.Lift.Instances
+    PlutusTx.Lift.TH
     PlutusTx.Lift.THUtils
 
   build-depends:
